# Simplify ascii string concat and skip some allocations