--- a/string_test.go
+++ b/string_test.go
@@ -12,3 +12,19 @@ func TestStringOOBProperties(t *testing.T) {
 
 	testScript1(SCRIPT, valueInt(1), t)
 }
+
+func BenchmarkASCIIConcat(b *testing.B) {
+	vm := New()
+
+	b.ResetTimer()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		_, err := vm.RunString(`{let result = "ab";
+		for (let i = 0 ; i < 10;i++) {
+			result += result;
+		}}`)
+		if err != nil {
+			b.Fatalf("Unexpected errors %s", err)
+		}
+	}
+}