With this simplified code it skips some allocations of byte slices

```
 benchstat old.bench new.bench
 name           old time/op    new time/op    delta
 ASCIIConcat-8    44.3µs ± 5%    40.2µs ±12%   -9.18%  (p=0.000 n=10+10)

 name           old alloc/op   new alloc/op   delta
 ASCIIConcat-8    14.0kB ± 0%     9.9kB ± 0%  -29.28%  (p=0.000 n=10+10)

 name           old allocs/op  new allocs/op  delta
 ASCIIConcat-8       131 ± 0%       121 ± 0%   -7.63%  (p=0.000 n=10+10)
```